refactor: Refactor data handling modules for improved organization and naming conventions
